CLDR-14615 Modernize: move Info Panel code to cldrInfo.js
-Move some Info Panel functions from cldrTable to cldrInfo

-Move appendIcon to cldrDom; called from both cldrTable and cldrInfo

-Update imports and exports as needed

-Fix recently introduced bug in cldrGui: total not const
