+++ tools/cldr-apps/js/src/esm/cldrTable.js
@@ -393,7 +393,7 @@ function reallyUpdateRow(tr, theRow) {
    * Update the vote info.
    */
   if (theRow.voteResolver) {
-    updateRowVoteInfo(tr, theRow);
+    cldrInfo.updateRowVoteInfo(tr, theRow);
   } else {
     tr.voteDiv = null;
   }
@@ -664,378 +664,6 @@ function getRowApprovalStatusClass(theRow) {
     }
   }
   return statusClass;
-}
-
-/**
- * Update the vote info for this row.
- *
- * Set up the "vote div".
- *
- * @param tr the table row
- * @param theRow the data from the server for this row
- *
- * Called by updateRow.
- *
- * TODO: shorten this function by using subroutines.
- */
-function updateRowVoteInfo(tr, theRow) {
-  if (!theRow) {
-    console.error("theRow is null or undefined in updateRowVoteInfo");
-    return;
-  }
-  var vr = theRow.voteResolver;
-  tr.voteDiv = document.createElement("div");
-  tr.voteDiv.className = "voteDiv";
-  const surveyUser = cldrStatus.getSurveyUser();
-  if (theRow.voteVhash && theRow.voteVhash !== "" && surveyUser) {
-    var voteForItem = theRow.items[theRow.voteVhash];
-    if (
-      voteForItem &&
-      voteForItem.votes &&
-      voteForItem.votes[surveyUser.id] &&
-      voteForItem.votes[surveyUser.id].overridedVotes
-    ) {
-      tr.voteDiv.appendChild(
-        cldrDom.createChunk(
-          cldrText.sub("override_explain_msg", {
-            overrideVotes: voteForItem.votes[surveyUser.id].overridedVotes,
-            votes: surveyUser.votecount,
-          }),
-          "p",
-          "helpContent"
-        )
-      );
-    }
-    if (theRow.voteVhash !== theRow.winningVhash && theRow.canFlagOnLosing) {
-      if (!theRow.rowFlagged) {
-        cldrSurvey.addIcon(tr.voteDiv, "i-stop");
-        tr.voteDiv.appendChild(
-          cldrDom.createChunk(
-            cldrText.sub("mustflag_explain_msg", {}),
-            "p",
-            "helpContent"
-          )
-        );
-      } else {
-        cldrSurvey.addIcon(tr.voteDiv, "i-flag");
-        tr.voteDiv.appendChild(
-          cldrDom.createChunk(cldrText.get("flag_desc", "p", "helpContent"))
-        );
-      }
-    }
-  }
-  if (!theRow.rowFlagged && theRow.canFlagOnLosing) {
-    cldrSurvey.addIcon(tr.voteDiv, "i-flag-d");
-    tr.voteDiv.appendChild(
-      cldrDom.createChunk(cldrText.get("flag_d_desc", "p", "helpContent"))
-    );
-  }
-  /*
-   * The value_vote array has an even number of elements,
-   * like [value0, vote0, value1, vote1, value2, vote2, ...].
-   */
-  var n = 0;
-  while (n < vr.value_vote.length) {
-    var value = vr.value_vote[n++];
-    var vote = vr.value_vote[n++];
-    if (value == null /* TODO: impossible? */ || value === NO_WINNING_VALUE) {
-      continue;
-    }
-    var item = tr.rawValueToItem[value]; // backlink to specific item in hash
-    if (item == null) {
-      continue;
-    }
-    var vdiv = cldrDom.createChunk(
-      null,
-      "table",
-      "voteInfo_perValue table table-vote"
-    );
-    var valdiv = cldrDom.createChunk(
-      null,
-      "div",
-      n > 2 ? "value-div" : "value-div first"
-    );
-    // heading row
-    var vrow = cldrDom.createChunk(
-      null,
-      "tr",
-      "voteInfo_tr voteInfo_tr_heading"
-    );
-    if (
-      item.rawValue === cldrSurvey.INHERITANCE_MARKER ||
-      (item.votes && Object.keys(item.votes).length > 0)
-    ) {
-      vrow.appendChild(
-        cldrDom.createChunk(
-          cldrText.get("voteInfo_orgColumn"),
-          "td",
-          "voteInfo_orgColumn voteInfo_td"
-        )
-      );
-    }
-    var isection = cldrDom.createChunk(null, "div", "voteInfo_iconBar");
-    var isectionIsUsed = false;
-    var vvalue = cldrDom.createChunk(
-      "User",
-      "td",
-      "voteInfo_valueTitle voteInfo_td"
-    );
-    var vbadge = cldrDom.createChunk(vote, "span", "badge");
-
-    /*
-     * Note: we can't just check for item.pClass === "winner" here, since, for example, the winning value may
-     * have value = cldrSurvey.INHERITANCE_MARKER and item.pClass = "alias".
-     */
-    if (value === theRow.winningValue) {
-      const statusClass = getRowApprovalStatusClass(theRow);
-      const statusTitle = cldrText.get(statusClass);
-      appendIcon(
-        isection,
-        "voteInfo_winningItem d-dr-" + statusClass,
-        cldrText.sub("draftStatus", [statusTitle])
-      );
-      isectionIsUsed = true;
-    }
-    if (item.isBaselineValue) {
-      appendIcon(isection, "i-star", cldrText.get("voteInfo_baseline_desc"));
-      isectionIsUsed = true;
-    }
-    cldrSurvey.setLang(valdiv);
-    if (value === cldrSurvey.INHERITANCE_MARKER) {
-      /*
-       * theRow.inheritedValue can be undefined here; then do not append
-       */
-      if (theRow.inheritedValue) {
-        cldrVote.appendItem(valdiv, theRow.inheritedValue, item.pClass);
-        valdiv.appendChild(
-          cldrDom.createChunk(cldrText.get("voteInfo_votesForInheritance"), "p")
-        );
-      }
-    } else {
-      cldrVote.appendItem(
-        valdiv,
-        value,
-        value === theRow.winningValue ? "winner" : "value"
-      );
-      if (value === theRow.inheritedValue) {
-        valdiv.appendChild(
-          cldrDom.createChunk(
-            cldrText.get("voteInfo_votesForSpecificValue"),
-            "p"
-          )
-        );
-      }
-    }
-    if (isectionIsUsed) {
-      valdiv.appendChild(isection);
-    }
-    vrow.appendChild(vvalue);
-    var cell = cldrDom.createChunk(
-      null,
-      "td",
-      "voteInfo_voteTitle voteInfo_voteCount voteInfo_td" + ""
-    );
-    cell.appendChild(vbadge);
-    vrow.appendChild(cell);
-    vdiv.appendChild(vrow);
-    const itemVotesLength = item.votes ? Object.keys(item.votes).length : 0;
-    const anon =
-      itemVotesLength == 1 &&
-      item.votes[Object.keys(item.votes)[0]].level === "anonymous";
-    if (itemVotesLength == 0 || anon) {
-      var vrow = cldrDom.createChunk(
-        null,
-        "tr",
-        "voteInfo_tr voteInfo_orgHeading"
-      );
-      vrow.appendChild(
-        cldrDom.createChunk(
-          cldrText.get("voteInfo_noVotes"),
-          "td",
-          "voteInfo_noVotes voteInfo_td"
-        )
-      );
-      const anonVoter = anon ? cldrText.get("voteInfo_anon") : null;
-      vrow.appendChild(
-        cldrDom.createChunk(anonVoter, "td", "voteInfo_noVotes voteInfo_td")
-      );
-      vdiv.appendChild(vrow);
-    } else {
-      updateRowVoteInfoForAllOrgs(theRow, vr, value, item, vdiv);
-    }
-    tr.voteDiv.appendChild(valdiv);
-    tr.voteDiv.appendChild(vdiv);
-  }
-  if (vr.valueIsLocked) {
-    tr.voteDiv.appendChild(
-      cldrDom.createChunk(
-        cldrText.get("valueIsLocked"),
-        "p",
-        "alert alert-warning fix-popover-help"
-      )
-    );
-  } else if (vr.requiredVotes) {
-    var msg = cldrText.sub("explainRequiredVotes", {
-      requiredVotes: vr.requiredVotes,
-    });
-    tr.voteDiv.appendChild(
-      cldrDom.createChunk(msg, "p", "alert alert-warning fix-popover-help")
-    );
-  }
-  // done with voteresolver table
-  if (CLDR_TABLE_DEBUG) {
-    tr.voteDiv.appendChild(cldrDom.createChunk(vr.raw, "p", "debugStuff"));
-  }
-}
-
-/**
- * Update the vote info for one candidate item in this row, looping through all the orgs.
- * Information will be displayed in the Information Panel (right edge of window).
- *
- * @param theRow the row
- * @param vr the vote resolver
- * @param value the value of the candidate item
- * @param item the candidate item
- * @param vdiv a table created by the caller as vdiv = cldrDom.createChunk(null, "table", "voteInfo_perValue table table-vote")
- */
-function updateRowVoteInfoForAllOrgs(theRow, vr, value, item, vdiv) {
-  for (let org in vr.orgs) {
-    var theOrg = vr.orgs[org];
-    var vrRaw = {};
-    var orgVoteValue = theOrg.votes[value];
-    /*
-     * We should display something under "Org." and "User" even when orgVoteValue is zero (not undefined),
-     * for "anonymous" imported losing votes. Therefore do not require orgVoteValue > 0 here.
-     * There does not appear to be any circumstance where we need to hide a zero vote count (on the client).
-     * If we do discover such a circumstance, we could display 0 vote only if voter is "anonymous";
-     * currently such voters have org = "cldr"; but if we don't need such a dependency here, don't add it.
-     * Reference: https://unicode.org/cldr/trac/ticket/11517
-     */
-    if (orgVoteValue !== undefined) {
-      // someone in the org actually voted for it
-      var topVoter = null; // top voter for this item
-      var orgsVote = theOrg.orgVote == value; // boolean
-      var topVoterTime = 0; // Calculating the latest time for a user from same org
-      if (orgsVote) {
-        // find a top-ranking voter to use for the top line
-        for (var voter in item.votes) {
-          if (
-            item.votes[voter].org == org &&
-            item.votes[voter].votes == theOrg.votes[value]
-          ) {
-            if (topVoterTime != 0) {
-              // Get the latest time vote only
-              if (vr.nameTime[`#${topVoter}`] < vr.nameTime[`#${voter}`]) {
-                topVoter = voter;
-                topVoterTime = vr.nameTime[`#${topVoter}`];
-              }
-            } else {
-              topVoter = voter;
-              topVoterTime = vr.nameTime[`#${topVoter}`];
-            }
-          }
-        }
-      } else {
-        // just find someone in the right org..
-        for (var voter in item.votes) {
-          if (item.votes[voter].org == org) {
-            topVoter = voter;
-            break;
-          }
-        }
-      }
-      // ORG SUBHEADING row
-
-      /*
-       * This only affects cells ("td" elements) with style "voteInfo_voteCount", which appear in the info panel,
-       * and which have contents like '<span class="badge">12</span>'. If the "fallback" style is added, then
-       * these circled numbers are surrounded (outside the circle) by a colored background.
-       *
-       * TODO: see whether the colored background is actually wanted in this context, around the numbers.
-       * For now, display it, and use item.pClass rather than literal "fallback" so the color matches when
-       * item.pClass is "alias", "fallback_root", etc.
-       */
-      var baileyClass =
-        item.rawValue === cldrSurvey.INHERITANCE_MARKER
-          ? " " + item.pClass
-          : "";
-      var vrow = cldrDom.createChunk(
-        null,
-        "tr",
-        "voteInfo_tr voteInfo_orgHeading"
-      );
-      vrow.appendChild(
-        cldrDom.createChunk(org, "td", "voteInfo_orgColumn voteInfo_td")
-      );
-      if (item.votes[topVoter]) {
-        vrow.appendChild(createVoter(item.votes[topVoter])); // voteInfo_td
-      } else {
-        vrow.appendChild(createVoter(null));
-      }
-      if (orgsVote) {
-        var cell = cldrDom.createChunk(
-          null,
-          "td",
-          "voteInfo_orgsVote voteInfo_voteCount voteInfo_td" + baileyClass
-        );
-        cell.appendChild(cldrDom.createChunk(orgVoteValue, "span", "badge"));
-        vrow.appendChild(cell);
-      } else {
-        vrow.appendChild(
-          cldrDom.createChunk(
-            orgVoteValue,
-            "td",
-            "voteInfo_orgsNonVote voteInfo_voteCount voteInfo_td" + baileyClass
-          )
-        );
-      }
-      vdiv.appendChild(vrow);
-      // now, other rows:
-      for (var voter in item.votes) {
-        if (
-          item.votes[voter].org != org || // wrong org or
-          voter == topVoter
-        ) {
-          // already done
-          continue; // skip
-        }
-        // OTHER VOTER row
-        var vrow = cldrDom.createChunk(null, "tr", "voteInfo_tr");
-        vrow.appendChild(
-          cldrDom.createChunk("", "td", "voteInfo_orgColumn voteInfo_td")
-        ); // spacer
-        vrow.appendChild(createVoter(item.votes[voter])); // voteInfo_td
-        vrow.appendChild(
-          cldrDom.createChunk(
-            item.votes[voter].votes,
-            "td",
-            "voteInfo_orgsNonVote voteInfo_voteCount voteInfo_td" + baileyClass
-          )
-        );
-        vdiv.appendChild(vrow);
-      }
-    }
-  }
-}
-
-/**
- * Create an element representing a voter, including a link to the voter's email
- *
- * @param v the voter
- * @return the element
- */
-function createVoter(v) {
-  if (v == null) {
-    return cldrDom.createChunk("(missing information)!", "i", "stopText");
-  }
-  var div = cldrDom.createChunk(
-    v.name || cldrText.get("emailHidden"),
-    "td",
-    "voteInfo_voterInfo voteInfo_td"
-  );
-  div.setAttribute("data-name", v.name || cldrText.get("emailHidden"));
-  div.setAttribute("data-email", v.email || "");
-  return div;
 }
 
 /*
@@ -1402,7 +1030,11 @@ function addVitem(td, tr, theRow, item, newButton) {
   checkLRmarker(choiceField, item.value);
 
   if (item.isBaselineValue == true) {
-    appendIcon(choiceField, "i-star", cldrText.get("voteInfo_baseline_desc"));
+    cldrDom.appendIcon(
+      choiceField,
+      "i-star",
+      cldrText.get("voteInfo_baseline_desc")
+    );
   }
   if (item.votes && !isWinner) {
     if (
@@ -1445,7 +1077,7 @@ function addVitem(td, tr, theRow, item, newButton) {
   div.appendChild(choiceField);
 
   // wire up the onclick function for the Info Panel
-  td.showFn = item.showFn = showItemInfoFn(theRow, item);
+  td.showFn = item.showFn = cldrInfo.showItemInfoFn(theRow, item);
   div.popParent = tr;
   cldrInfo.listen(null, tr, div, td.showFn);
   td.appendChild(div);
@@ -1453,13 +1085,6 @@ function addVitem(td, tr, theRow, item, newButton) {
   if (item.example && item.value != item.examples) {
     appendExample(div, item.example);
   }
-}
-
-function appendIcon(toElement, className, title) {
-  var e = cldrDom.createChunk(null, "div", className);
-  e.title = title;
-  toElement.appendChild(e);
-  return e;
 }
 
 /**
@@ -1488,120 +1113,6 @@ function appendExample(parent, text, loc) {
   cldrSurvey.setLang(div, loc);
   parent.appendChild(div);
   return div;
-}
-
-/**
- * Return a function that will show info for the given item in the Info Panel.
- *
- * @param theRow the data row
- * @param item the candidate item
- * @returns the function
- *
- * Called only by addVitem.
- */
-function showItemInfoFn(theRow, item) {
-  return function (td) {
-    var h3 = document.createElement("div");
-    var displayValue = item.value;
-    if (item.value === cldrSurvey.INHERITANCE_MARKER) {
-      displayValue = theRow.inheritedValue;
-    }
-
-    var span = cldrVote.appendItem(h3, displayValue, item.pClass);
-    cldrSurvey.setLang(span);
-    h3.className = "span";
-    td.appendChild(h3);
-
-    if (item.value) {
-      /*
-       * Strings produced here, used as keys for cldrText.sub(), may include:
-       *  "pClass_winner", "pClass_alias", "pClass_fallback", "pClass_fallback_code", "pClass_fallback_root", "pClass_loser".
-       *  See getPClass in DataSection.java.
-       *
-       *  TODO: why not show stars, etc., here?
-       */
-      h3.appendChild(
-        cldrDom.createChunk(
-          cldrText.sub("pClass_" + item.pClass, item),
-          "p",
-          "pClassExplain"
-        )
-      );
-    }
-
-    if (item.value === cldrSurvey.INHERITANCE_MARKER) {
-      addJumpToOriginal(theRow, h3);
-    }
-
-    var newDiv = document.createElement("div");
-    td.appendChild(newDiv);
-
-    if (item.tests) {
-      newDiv.innerHTML = cldrSurvey.testsToHtml(item.tests);
-    } else {
-      newDiv.innerHTML = "<i>no tests</i>";
-    }
-
-    if (item.example) {
-      appendExample(td, item.example);
-    }
-  }; // end function(td)
-}
-
-/**
- * Add a link in the Info Panel for "Jump to Original" (cldrText.get('followAlias')),
- * if theRow.inheritedLocale or theRow.inheritedXpid is defined.
- *
- * Normally at least one of theRow.inheritedLocale and theRow.inheritedXpid should be
- * defined whenever we have an INHERITANCE_MARKER item. Otherwise an error is reported
- * by checkRowConsistency.
- *
- * This is currently (2018-12-01) the only place inheritedLocale or inheritedXpid is used on the client.
- * An alternative would be for the server to send the link (clickyLink.href), instead of inheritedLocale
- * and inheritedXpid, to the client, avoiding the need for the client to know so much, including the need
- * to replace 'code-fallback' with 'root' or when to use cldrStatus.getCurrentLocale() in place of inheritedLocale
- * or use xpstrid in place of inheritedXpid.
- *
- * @param theRow the row
- * @param el the element to which to append the link
- */
-function addJumpToOriginal(theRow, el) {
-  if (theRow.inheritedLocale || theRow.inheritedXpid) {
-    var loc = theRow.inheritedLocale;
-    var xpstrid = theRow.inheritedXpid || theRow.xpstrid;
-    if (!loc) {
-      loc = cldrStatus.getCurrentLocale();
-    } else if (loc === "code-fallback") {
-      /*
-       * Never use 'code-fallback' in the link, use 'root' instead.
-       * On the server, 'code-fallback' sometimes goes by the name XMLSource.CODE_FALLBACK_ID.
-       * Reference: https://unicode.org/cldr/trac/ticket/11622
-       */
-      loc = "root";
-    }
-    if (
-      xpstrid === theRow.xpstrid && // current hash
-      loc === cldrStatus.getCurrentLocale()
-    ) {
-      // current locale
-      // i.e., following the alias would come back to the current item
-      el.appendChild(
-        cldrDom.createChunk(
-          cldrText.get("noFollowAlias"),
-          "span",
-          "followAlias"
-        )
-      );
-    } else {
-      var clickyLink = cldrDom.createChunk(
-        cldrText.get("followAlias"),
-        "a",
-        "followAlias"
-      );
-      clickyLink.href = "#/" + loc + "//" + xpstrid;
-      el.appendChild(clickyLink);
-    }
-  }
 }
 
 /**
@@ -1759,6 +1270,9 @@ function getValidWinningValue(theRow) {
 }
 
 export {
+  NO_WINNING_VALUE,
+  appendExample,
+  getRowApprovalStatusClass,
   insertRows,
   updateRow,
   /*
